+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2022-09-28
+### Added
+- Warmup-1 intMax
+
 ## [1.3.0] - 2022-09-27
 ### Added
 - Warmup-1 delDel